+++ allensdk/brain_observatory/nwb/behavior_ophys_nwb_extension_builder.py
@@ -4,7 +4,7 @@ from allensdk.brain_observatory.behavior.schemas import (
     BehaviorMetadataSchema,
     OphysBehaviorMetadataSchema,
     BehaviorTaskParametersSchema,
-    SubjectMetadataSchema
+    SubjectMetadataSchema, OphysEyeTrackingRigMetadataSchema
 )
 from allensdk.brain_observatory.nwb.metadata import (
     create_pynwb_extension_from_schemas
@@ -15,7 +15,7 @@ if __name__ == "__main__":
     # Run this module to regenerate the extension yaml files into this dir:
     prefix = 'ndx-aibs-behavior-ophys'
     schemas = [BehaviorTaskParametersSchema, SubjectMetadataSchema,
-               BehaviorMetadataSchema, OphysBehaviorMetadataSchema]
+               BehaviorMetadataSchema, OphysBehaviorMetadataSchema, OphysEyeTrackingRigMetadataSchema]
     curr_dir = os.path.abspath(os.path.dirname(__file__))
     create_pynwb_extension_from_schemas(schemas, prefix, save_dir=curr_dir)
     print("Creation of NWB extension complete!")